Add size tracking for files and folders in MB
Introduced a `size` property to `File` and `Folder` classes for size tracking in megabytes. Updated constructors, services, and DTOs to support size calculations and propagation. Added recursive methods for folder size updates and recalculations. Updated migrations to move `Size` from `Metadata` to `File` and `Folder` entities. Enhanced UI to display file sizes and adjusted CSS for styling. Refactored sorting logic and removed redundant code. Updated tests and seed data to align with schema changes.

+++ CloudStoragePlatform.Core/DTO/FileOrFolderMetadataResponse.cs
@@ -27,6 +27,6 @@ namespace CloudStoragePlatform.Core.DTO
         public int OpenCount { get; set; }
 
         public int ShareCount { get; set; }
-        public long Size { get; set; }
+        public float Size { get; set; }
     }
 }
